% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/methods.R
\docType{methods}
\name{checkW4M}
\alias{checkW4M}
\alias{checkW4M,ExpressionSet-method}
\title{Checking the consistency of an ExpressionSet instance with W4M format}
\usage{
checkW4M(eset, ...)

\S4method{checkW4M}{ExpressionSet}(eset, ...)
}
\arguments{
\item{eset}{An S4 object of class ExpressionSet.}

\item{...}{Currently not used.}
}
\value{
Invisible TRUE logical in case of success (otherwise generates an
error)
}
\description{
Checking the consistency of an ExpressionSet instance with W4M format
}
\examples{
sacSet <- fromW4M(file.path(path.package("ropls"), "extdata"))
print(checkW4M(sacSet))
}
\author{
Etienne Thevenot, \email{etienne.thevenot@cea.fr}
}
