% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{coef,opls-method}
\alias{coef,opls-method}
\alias{coef.opls}
\title{Coefficients method for (O)PLS models}
\usage{
\S4method{coef}{opls}(object, ...)
}
\arguments{
\item{object}{An S4 object of class \code{opls}, created by \code{opls}
function.}

\item{...}{Currently not used.}
}
\value{
Numeric matrix of coefficients (number of rows equals the number of
variables, and the number of columns equals the number of responses)
}
\description{
Coefficients of the (O)PLS(-DA) regression model
}
\examples{

data(sacurine)
attach(sacurine)

sacurine.plsda <- opls(dataMatrix,
                       sampleMetadata[, "gender"])

head(coef(sacurine.plsda))

detach(sacurine)

}
\author{
Etienne Thevenot, \email{etienne.thevenot@cea.fr}
}
