% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cornell}
\alias{cornell}
\title{Octane of various blends of gasoline}
\format{A data frame with the following parameters:
\itemize{
\item num mixture number
\item x1 proportion of
component 1
\item x2 proportion of component 2
\item x3 proportion of component 3
\item x4 proportion of component 4
\item x5 proportion of component 5
\item x6 proportion of component 6
\item x7 proportion of component 7 Note: the 7 variables are correlated since they sum up to 1
\item y octane (quantitative variable)
}}
\source{
Tenenhaus (1998), Table 6, page 78.
}
\value{
Data frame (numeric type only; the first column can be transformed
into row names) with 12 rows and 9 columns corresponding to the 'num'ber of
the mixture (column 1), the proportion of each of the 7 'x' components
within the mixture (columns 2-8), and the octane indice 'y' (column 9). For
details see the 'Format' section above.
}
\description{
Twelve mixture component proportions of the blend are analysed
}
\references{
Tenenhaus (1998). La regression PLS: theorie et pratique. Paris:
Editions Technip.
}
\keyword{datasets}
