% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{fitted,opls-method}
\alias{fitted,opls-method}
\alias{fitted.opls}
\title{Fitted method for 'opls' objects}
\usage{
\S4method{fitted}{opls}(object, ...)
}
\arguments{
\item{object}{An S4 object of class \code{opls}, created by the \code{opls}
function.}

\item{...}{Currently not used.}
}
\value{
Predictions (either a vector, factor, or matrix depending on the y
response used for training the model)
}
\description{
Returns predictions of the (O)PLS(-DA) model on the training dataset
}
\examples{

data(sacurine)
attach(sacurine)
sacurine.plsda <- opls(dataMatrix, sampleMetadata[, "gender"])

fitted(sacurine.plsda)

detach(sacurine)

}
\author{
Etienne Thevenot, \email{etienne.thevenot@cea.fr}
}

