% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{foods}
\alias{foods}
\title{Food consumption patterns accross European countries (FOODS)}
\format{A data frame with the following parameters:
\itemize{
\item Country Name of the country
\item Gr_CoffeGround Coffee
\item Inst_Coffe Instant Coffee
\item Tea Tea \item Sweetner Sweetner
\item Biscuits Biscuits \item Pa_Soup Powder Soup
\item Ti_Soup Tin Soup \item In_Potat Instant Potatoes
\item Fro_Fish Frozen Fish
\item Fro_Veg Frozen Vegetables
\item Apples Apples
\item Oranges Oranges
\item Ti_Fruit Tin Fruit
\item Jam Jam
\item Garlic Garlic
\item Butter Butter
\item Margarine Margarine
\item Olive_Oil Olive Oil
\item Yoghurt Yoghurt
\item Crisp_Brea Crisp Bread
}}
\value{
Data frame (numeric type except the first column, which can be
transformed into row names) with 16 rows and 21 columns, corresponding to
the 'Country' (column 1), followed by the consumption of each of the 20 food
items (columns 2-21). For details see the 'Format' section above.
}
\description{
The relative consumption of 20 food items was compiled for 16 countries. The
values range between 0 and 100 percent and a high value corresponds to a
high consumption. The dataset contains 3 missing data.
}
\references{
Eriksson et al. (2006). Multi- and Megarvariate Data Analysis.
Umetrics Academy. pp.10, 33, 48.
}
\keyword{datasets}
