% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{imageF}
\alias{imageF}
\title{Image of a numerical matrix}
\usage{
imageF(matrixMN, paletteC = c("heat", "revHeat", "grey", "revGrey",
  "palette", "ramp")[1], mainC = NA, logL = FALSE, fig.pdfC = NA)
}
\arguments{
\item{matrixMN}{numerical matrix}

\item{paletteC}{color palette to be used (either 'heat' -default, 'revHeat',
'grey', 'revGrey', 'palette', or 'ramp')}

\item{mainC}{title (by default, the name of the matrixMN variable will be used)}

\item{logL}{should the matrix values be log transformed?}

\item{fig.pdfC}{file name for the figure (with '.pdf' extension), if set to
NA (default), the figure is displayed on the screen}
}
\value{
No output.
}
\description{
Visualization of a numerical matrix
}
\examples{
data(sacurine)
ropls::imageF(sacurine[['dataMatrix']])
}
