% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{NCI60}
\alias{NCI60}
\title{Microarray gene expression profiles of the NCI 60 cell lines from 4 different platforms}
\format{
A list with the following elements:
\itemize{
\item mae dataset in the MultiAssayExperiment format
\item mds dataset in the MultiDataSet format
}
}
\value{
List containing the NCI60_4arrays dataset from the omicade4 package in the
MultiAssayExperiment and MultiDataSet formats.
}
\description{
The 'NCI60_4arrays' dataset from the 'omicade4' package is provided here in
the MultiAssayExperiment and MultiDataSet formats. The description of the dataset provided
by the omicade4 package is as follows: 'The 60 human tumour cell lines are derived from patients with leukaemia, melanoma,
lung, colon, central nervous system, ovarian, renal, breast and prostate cancers.
The cell line panel is widely used in anti-cancer drug screen. In this dataset,
a subset of microarray gene expression of the NCI 60 cell lines from four different platforms
are provided.'
}
\references{
Reinhold WC, Sunshine M, Liu H, Varma S, Kohn KW, Morris J, Doroshow J,
Pommier Y (2012). CellMiner: A Web-Based Suite of Genomic and Pharmacologic Tools to
Explore Transcript and Drug Patterns in the NCI-60 Cell Line Set. Cancer Research.
DOI: 10.1158/0008-5472.CAN-12-1370
}
\keyword{datasets}
