% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{oplsMultiDataSet-class}
\alias{oplsMultiDataSet-class}
\title{Class "oplsMultiDataSet"}
\description{
An S4 class to store PCA and (O)PLS(-DA) models generated by the application
of \code{opls} to a MultiDataSet
}
\section{Slots}{

\describe{
\item{\code{oplsLs}}{List: List of instances from the 'opls' class corresponding to
the models built on each ExpresssionSet}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("oplsMultiDataSet", ...)} or by applying the \code{opls} function to
a MultiDataSet instance
}

\examples{
In progress
}
\seealso{
\code{\link{opls}}
}
\author{
Etienne Thevenot, \email{etienne.thevenot@cea.fr}
}
