% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{print,opls-method}
\alias{print,opls-method}
\alias{print.opls}
\title{Print method for 'opls' objects}
\usage{
\S4method{print}{opls}(x, ...)
}
\arguments{
\item{x}{An S4 object of class \code{opls}, created by the \code{opls}
function.}

\item{...}{Currently not used.}
}
\value{
Invisible.
}
\description{
Displays information about the dataset and the model.
}
\examples{

data(sacurine)
attach(sacurine)
sacurine.plsda <- opls(dataMatrix, sampleMetadata[, "gender"])

print(sacurine.plsda)

detach(sacurine)

}
\author{
Etienne Thevenot, \email{etienne.thevenot@cea.fr}
}

