% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/helpers.R
\name{checkW4M}
\alias{checkW4M}
\alias{checkW4M,SummarizedExperiment-method}
\alias{checkW4M,ExpressionSet-method}
\title{Checking the consistency of a SummarizedExperiment or ExpressionSet instance with W4M format}
\usage{
checkW4M(x)

\S4method{checkW4M}{SummarizedExperiment}(x)

\S4method{checkW4M}{ExpressionSet}(x)
}
\arguments{
\item{x}{An S4 object of class \code{SummarizedExperiment} or \code{ExpressionSet}}
}
\value{
Invisible TRUE logical in case of success (otherwise generates an
error)
}
\description{
Checking the consistency of a SummarizedExperiment or ExpressionSet instance with W4M format
}
\examples{
sacSet <- fromW4M(file.path(path.package("ropls"), "extdata"))
print(checkW4M(sacSet))
}
\author{
Etienne Thevenot, \email{etienne.thevenot@cea.fr}
}
