% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/methods.R
\docType{methods}
\name{getMset}
\alias{getMset}
\alias{getMset,oplsMultiDataSet-method}
\alias{getMset,}
\alias{oplsMultiDataSet-method}
\title{getMset method}
\usage{
getMset(object, ...)

\S4method{getMset}{oplsMultiDataSet}(object)
}
\arguments{
\item{object}{An S4 object of class \code{oplsMultiDataSet}, created by \code{opls}
function applied to a MultiDataSet}

\item{...}{Currently not used}
}
\value{
An S4 object of class \code{MultiDataSet}.
}
\description{
Extracts the complemented MultiDataSet when opls has been applied to a MultiDataSet
}
\examples{
library(brgedata)
data("brge_gexp") # Get the ExpressionSet from brgedata 
gSet1 <- brge_gexp[sample(1:21916, 300), 1:50]   # Create ExpressionSet with samples from 01 to 50
gSet2 <- brge_gexp[sample(1:21916, 300), 35:61]  # Create ExpressionSet with samples from 35 to 64
brgeMset <- MultiDataSet::createMultiDataSet()
brgeMset <- MultiDataSet::add_eset(brgeMset, gSet1, dataset.type = "expression",
dataset.name = "gSet1", warnings = FALSE)
brgeMset <- MultiDataSet::add_eset(brgeMset, gSet2, dataset.type = "expression",
dataset.name = "gSet2", warnings = FALSE)
brgeMset
brgePca <- ropls::opls(brgeMset)
brgeMset <- ropls::getMset(brgePca)

}
