% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/methods.R
\docType{methods}
\name{getPcaVarVn}
\alias{getPcaVarVn}
\alias{getPcaVarVn,opls-method}
\alias{getPcaVarVn,opls-method}
\title{getPcaVarVn method for PCA models}
\usage{
getPcaVarVn(object, ...)

\S4method{getPcaVarVn}{opls}(object)
}
\arguments{
\item{object}{An S4 object of class \code{opls}, created by \code{opls}
function.}

\item{...}{Currently not used.}
}
\value{
Numeric vector with the same length as the number of components
}
\description{
Variance of the components (score vectors)
}
\examples{

data(sacurine)
attach(sacurine)

sacurine.pca <- opls(dataMatrix)

getPcaVarVn(sacurine.pca)

detach(sacurine)

}
\author{
Etienne Thevenot, \email{etienne.thevenot@cea.fr}
}
