% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/methods.R
\docType{methods}
\name{getSummaryDF}
\alias{getSummaryDF}
\alias{getSummaryDF,opls-method}
\alias{getSummaryDF,opls-method}
\title{getSummaryDF method for PCA/(O)PLS models}
\usage{
getSummaryDF(object, ...)

\S4method{getSummaryDF}{opls}(object)
}
\arguments{
\item{object}{An S4 object of class \code{opls}, created by \code{opls}
function.}

\item{...}{Currently not used.}
}
\value{
Data frame
}
\description{
Summary of model metrics
}
\examples{

data(sacurine)
attach(sacurine)

sacurine.plsda <- opls(dataMatrix,
                       sampleMetadata[, "gender"])

getSummaryDF(sacurine.plsda)

detach(sacurine)

}
\author{
Etienne Thevenot, \email{etienne.thevenot@cea.fr}
}
