% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{imageF}
\alias{imageF}
\title{Image of a numerical matrix}
\usage{
imageF(matrixMN, paletteC = c("heat", "revHeat", "grey", "revGrey",
  "palette", "ramp")[1], mainC = NA, logC = c("none", "log2",
  "log10")[1])
}
\arguments{
\item{matrixMN}{numerical matrix}

\item{paletteC}{color palette to be used (either 'heat', 'revHeat', 'grey',
'revGrey', 'palette', 'ramp')}

\item{mainC}{title (by default, the name of the matrixMN variable will be used)}

\item{logC}{optional log transformation of the non negative values from
matrixMN before plotting}
}
\value{
No output.
}
\description{
Visualization of a numerical matrix
}
\examples{
data(sacurine)
ropls::imageF(sacurine[['dataMatrix']])
}
