% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{linnerud}
\alias{linnerud}
\title{Linnerud Dataset}
\format{A data frame with the following parameters:
\itemize{
\item num subject number
\item weight weight
\item waist waist
\item pulse pulse
\item pullUp pull-up
\item squat situp
\item jump jump
}}
\source{
'mixOmics' 'linnerud' dataset.
}
\value{
Data frame (numeric type only; the first column can be transformed
into row names) with 20 rows and 7 columns corresponding to the subject's
'num'ber (column 1), the 3 physiological variables (columns 2-4), and the 3
exercise variables (columns 5-7). For details see the 'Format' section
above.
}
\description{
Three physiological and three exercise variables are measured on twenty
middle-aged men in a fitness club.
}
\references{
Tenenhaus (1998). La regression PLS: theorie et pratique. Paris:
Editions Technip.
}
\keyword{datasets}
