% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mark}
\alias{mark}
\title{'mark' Dataset}
\format{A data frame with the following parameters:
\itemize{
\item nom names of the students
\item math marks in mathematics
\item phys marks in physics
\item fran marks in french
\item angl marks in english
}}
\source{
'mark' dataset.
}
\value{
Data frame (numeric type except the first column, which can be
transformed into row names) with 9 rows and 5 columns, corresponding to the
name of the students (column 1), followed by the marks obtained in Maths,
Physics, French and English (columns 2-5). For details see the 'Format'
section above.
}
\description{
Examination marks obtained by French students in Mathematics, Physics,
French and English
}
\references{
Baccini (2010). Statistique Descriptive Multidimensionnelle
(pour les nuls).
}
\keyword{datasets}

