% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{strF}
\alias{strF}
\title{Printed summary of an R object}
\usage{
strF(inpMF, borderN = 2, bigMarkC = ",")
}
\arguments{
\item{inpMF}{Input matrix, dataframe or vector}

\item{borderN}{Number of border (first and last) rows and columns to display}

\item{bigMarkC}{Big mark separator for summary results}
}
\value{
This function has no output.
}
\description{
Displays the class, mode, size and first...last values of the object
}
\examples{

data(sacurine)
strF(sacurine[['dataMatrix']])
strF(sacurine[['sampleMetadata']])

}
\seealso{
\code{\link{str}}
}
\author{
Etienne Thevenot (CEA)
}
